added styles, a layout, and a dummy index page
I added a Bootstrap-based theme and stylesheet for the Web site,
as well as the beginnings of a layout template and a placeholder
index page.

Note that the use of CC-BY-SA is not 100% final, we must have a
LUG discussion in order to determine whether to add an -NC
option or not.