reconfigured the blog URL's to not be timezone dependent
Apparently, when you use DateDetailView, the date used for selecting
the object is interpreted as being in the active timezone. This
means that the URL's are dependent on the user's time zone. I
adjusted thte urlconf and views to avoid using it in DateDetailView.
(It's still probably useful to use the other date-based views for
browsing through the archives though.)